backend/requirements.txt
django-cors-headers==4.4.0
djangorestframework==3.15.2
pip-chill==1.0.3
psycopg2-binary==2.9.9
